+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/HttpLookupConfig.java
@@ -3,6 +3,7 @@ package com.getindata.connectors.http.internal.table.lookup;
 import java.io.Serializable;
 import java.util.Collections;
 import java.util.List;
+import java.util.Properties;
 
 import lombok.Builder;
 import lombok.Data;
@@ -20,4 +21,7 @@ public class HttpLookupConfig implements Serializable {
 
     @Builder.Default
     private final boolean useAsync = false;
+
+    @Builder.Default
+    private final Properties properties = new Properties();
 }