+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/HttpLookupTableSourceFactory.java
@@ -2,16 +2,14 @@ package com.getindata.connectors.http.internal.table.lookup;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
+import java.util.Properties;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import org.apache.flink.api.common.serialization.DeserializationSchema;
 import org.apache.flink.configuration.ConfigOption;
-import org.apache.flink.configuration.ConfigOptions;
-import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.DataTypes.Field;
@@ -28,6 +26,8 @@ import static org.apache.flink.table.api.DataTypes.FIELD;
 import static org.apache.flink.table.types.utils.DataTypeUtils.removeTimeAttribute;
 
 import com.getindata.connectors.http.internal.PollingClientFactory;
+import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
+import com.getindata.connectors.http.internal.utils.ConfigUtils;
 import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.ASYNC_POLLING;
 import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.URL;
 import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.URL_ARGS;
@@ -49,12 +49,11 @@ public class HttpLookupTableSourceFactory implements DynamicTableSourceFactory {
 
     @Override
     public DynamicTableSource createDynamicTableSource(Context context) {
-        final FactoryUtil.TableFactoryHelper helper =
+        FactoryUtil.TableFactoryHelper helper =
             FactoryUtil.createTableFactoryHelper(this, context);
 
         ReadableConfig readableConfig = helper.getOptions();
-
-        validateOptions(context, readableConfig);
+        helper.validateExcept(HttpConnectorConfigConstants.GID_CONNECTOR_HTTP);
 
         DecodingFormat<DeserializationSchema<RowData>> decodingFormat =
             helper.discoverDecodingFormat(
@@ -62,8 +61,8 @@ public class HttpLookupTableSourceFactory implements DynamicTableSourceFactory {
                 FactoryUtil.FORMAT
             );
 
-        PollingClientFactory<RowData> pollingClientFactory = new RestTablePollingClientFactory();
-        HttpLookupConfig lookupConfig = getHttpLookupOptions(readableConfig);
+        PollingClientFactory<RowData> pollingClientFactory = new JavaNetHttpPollingClientFactory();
+        HttpLookupConfig lookupConfig = getHttpLookupOptions(context, readableConfig);
 
         ResolvedSchema resolvedSchema = context.getCatalogTable().getResolvedSchema();
 
@@ -93,30 +92,17 @@ public class HttpLookupTableSourceFactory implements DynamicTableSourceFactory {
         return Set.of(URL_ARGS, ASYNC_POLLING);
     }
 
-    private HttpLookupConfig getHttpLookupOptions(ReadableConfig config) {
+    private HttpLookupConfig getHttpLookupOptions(Context context, ReadableConfig config) {
+
+        Properties httpConnectorProperties =
+            ConfigUtils.getHttpConnectorProperties(context.getCatalogTable().getOptions());
+
         return HttpLookupConfig.builder()
             .url(config.get(URL))
             .arguments(convertArguments(config.get((URL_ARGS))))
             .useAsync(config.get(ASYNC_POLLING))
+            .properties(httpConnectorProperties)
             .build();
-    }
-
-    // TODO FIX There is something ugly here. Verify with DataGenTableSourceFactory and refactor.
-    private void validateOptions(Context context, ReadableConfig readableConfig) {
-
-        Set<ConfigOption<?>> allOptions = new HashSet<>();
-        allOptions.addAll(optionalOptions());
-        allOptions.addAll(requiredOptions());
-        allOptions.add(ConfigOptions.key("connector").stringType().noDefaultValue());
-
-        FactoryUtil.validateFactoryOptions(requiredOptions(), allOptions, readableConfig);
-
-        Configuration options = new Configuration();
-        context.getCatalogTable().getOptions().forEach(options::setString);
-        Set<String> consumedOptionKeys = new HashSet<>();
-        allOptions.stream().map(ConfigOption::key).forEach(consumedOptionKeys::add);
-        FactoryUtil.validateUnconsumedKeys(factoryIdentifier(), options.keySet(),
-            consumedOptionKeys);
     }
 
     private List<String> convertArguments(String arguments) {